Move bake_exercises function into v1 class
